+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [NEXT]
 - Adds `MatrixMapGenerator`. Class that can help you create a map by double matrix.
 - Adds `TerrainBuilder`. Class that can help you create a map by double matrix with Sprite. [SpriteSheetModel](https://github.com/RafaelBarbosatec/bonfire/blob/master/example/assets/images/tile_random/earth_to_grass.png)
+- Adds `enabledDiagonalMovements` param in `MovementByJoystick` to control if you want diagonal movements.
 
 # [2.4.4]
 - Adds `scene` support. Now you can create a "cutscene" of easy way.